+++ src/routes/ordrer/fortryd/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import CancellationRequestDialog from '$lib/components/cancellation-request-dialog.svelte'
+  import CancelRequestForm from '$lib/components/cancel-request-form.svelte'
   import type { PageProps } from './$types'
 
   const { data, form }: PageProps = $props()
@@ -9,4 +9,9 @@
   <title>Fortryd køb | Vermouth.nu</title>
 </svelte:head>
 
-<CancellationRequestDialog actionData={form} mode="page" prefill={data.prefill} />
+<section class="px-11 py-16">
+  <div class="mx-auto max-w-2xl text-left">
+    <h1 class="mb-6 text-lg font-bold">Fortryd køb</h1>
+    <CancelRequestForm actionData={form} locale={data.locale} prefill={data.prefill} />
+  </div>
+</section>
