+++ src/routes/ordrer/fortryd/+page.ts
@@ -1,7 +1,10 @@
 import type { PageLoad } from './$types'
 
-export const load: PageLoad = ({ url }) => {
+export const load: PageLoad = async ({ parent, url }) => {
+  const { locale } = await parent()
+
   return {
+    locale,
     prefill: {
       customerName: url.searchParams.get('customerName') ?? '',
       email: url.searchParams.get('email') ?? '',
